requirements.txt
requests
gql
python-dotenv
websockets
web3
py-solc-x
